refactor(auth): move UserRepository interface to application/usecases
CLAUDE.md DDD-гейт: repository interfaces в пакете-потребителе
(usecase/), НЕ в domain/. C2 из 6-step ADR-6 sequence.

Interface relocates from `internal/modules/auth/domain/repositories/`
к `internal/modules/auth/application/usecases/`. 12 файлов обновили
импорты (sealed callers внутри auth/usecases используют bare name;
external callers — `usecases.UserRepository` alias).

Cross-module users module: `authRepos.UserRepository` →
`authUsecases.UserRepository` (temporary cross-module import до C7
narrow-port closure).

Build / test / lint зелёные после каждого commit'а.
