+++ internal/modules/auth/interfaces/http/handlers/auth_handler_unit_test.go
@@ -19,7 +19,6 @@ import (
 	"github.com/inf-sys-secretary-methodologist/inf-sys-secretary-methodist/internal/modules/auth/application/usecases"
 	"github.com/inf-sys-secretary-methodologist/inf-sys-secretary-methodist/internal/modules/auth/domain"
 	authEntities "github.com/inf-sys-secretary-methodologist/inf-sys-secretary-methodist/internal/modules/auth/domain/entities"
-	"github.com/inf-sys-secretary-methodologist/inf-sys-secretary-methodist/internal/modules/auth/domain/repositories"
 	emailServices "github.com/inf-sys-secretary-methodologist/inf-sys-secretary-methodist/internal/modules/notifications/domain/services"
 	"github.com/inf-sys-secretary-methodologist/inf-sys-secretary-methodist/internal/shared/security/totp"
 )
@@ -28,8 +27,8 @@ func init() {
 	gin.SetMode(gin.TestMode)
 }
 
-// Ensure MockUserRepository implements repositories.UserRepository
-var _ repositories.UserRepository = (*MockUserRepository)(nil)
+// Ensure MockUserRepository implements usecases.UserRepository
+var _ usecases.UserRepository = (*MockUserRepository)(nil)
 
 // MockUserRepository is a mock implementation of UserRepository for testing
 type MockUserRepository struct {
